Lots of changes
Added GI to Pause Scenes
Added TAA to every scene
Removed unused scenes
Disabled shadows on some lights for performance
Change the fridge on every scene to the new version
Changed the color of Stars not Won
Change the layout of the buttons in the main scene.
Changed the positions of the stars to above the buttons.

Computed every scene's  GI
Built Apk